pyproject.toml
[tool.poetry]
name = "structured-output"
version = "0.1.0"
description = ""
authors = ["Ethan Steininger <esteininger21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
uvicorn = "^0.28.0"
fastapi = "^0.110.0"
unstructured = {extras = ["pdf"], version = "^0.12.6"}
pydantic = {extras = ["email"], version = "^2.7.1"}
openai = "^1.23.6"
instructor = "^1.2.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
